Gui: split selection Python adapter and support headers
