+++ src/Gui/Selection/Selection.h
@@ -34,10 +34,12 @@
 #include <App/DocumentObserver.h>
 #include <Base/Observer.h>
 
+#include "SelectionChanges.h"
+#include "SelectionGate.h"
 #include "SelectionObject.h"
+#include "SelectionObserver.h"
+#include "SelectionTypes.h"
 
-
-using PyObject = struct _object;
 
 namespace App
 {
@@ -49,260 +51,8 @@ class PropertyLinkSubList;
 namespace Gui
 {
 
-enum class ResolveMode
-{
-    /// No resolve
-    NoResolve,
-    /// Sub-object with old style element name
-    OldStyleElement,
-    /// Sub-object with new style element name
-    NewStyleElement,
-    /// Resolve sub-object and follow link
-    FollowLink
-};
-
 class SelectionObject;
 class SelectionFilter;
-
-/** Transport the changes of the Selection
- * This class transports closer information what was changed in the
- * selection. It's an optional information and not all commands set this
- * information. If not set all observer of the selection assume a full change
- * and update everything (e.g 3D view). This is not a very good idea if, e.g. only
- * a small parameter has changed. Therefore one can use this class and make the
- * update of the document much faster!
- * @see Base::Observer
- */
-class GuiExport SelectionChanges
-{
-public:
-    enum MsgType
-    {
-        AddSelection,
-        RmvSelection,
-        SetSelection,
-        ClrSelection,
-        SetPreselect,  // to signal observer the preselect has changed
-        RmvPreselect,
-        SetPreselectSignal,  // to request 3D view to change preselect
-        PickedListChanged,
-        ShowSelection,       // to show a selection
-        HideSelection,       // to hide a selection
-        RmvPreselectSignal,  // to request 3D view to remove preselect
-        MovePreselect,       // to signal observer the mouse movement when preselect
-    };
-    enum class MsgSource
-    {
-        Any = 0,
-        Internal = 1,
-        TreeView = 2
-    };
-
-    SelectionChanges(
-        MsgType type = ClrSelection,
-        const char* docName = nullptr,
-        const char* objName = nullptr,
-        const char* subName = nullptr,
-        const char* typeName = nullptr,
-        float x = 0,
-        float y = 0,
-        float z = 0,
-        MsgSource subtype = MsgSource::Any
-    )
-        : Type(type)
-        , SubType(subtype)
-        , x(x)
-        , y(y)
-        , z(z)
-        , Object(docName, objName, subName)
-    {
-        pDocName = Object.getDocumentName().c_str();
-        pObjectName = Object.getObjectName().c_str();
-        pSubName = Object.getSubName().c_str();
-        if (typeName) {
-            TypeName = typeName;
-        }
-        pTypeName = TypeName.c_str();
-    }  // explicit bombs
-
-    SelectionChanges(
-        MsgType type,
-        const std::string& docName,
-        const std::string& objName,
-        const std::string& subName,
-        const std::string& typeName = std::string(),
-        float x = 0,
-        float y = 0,
-        float z = 0,
-        MsgSource subtype = MsgSource::Any
-    )
-        : Type(type)
-        , SubType(subtype)
-        , x(x)
-        , y(y)
-        , z(z)
-        , Object(docName.c_str(), objName.c_str(), subName.c_str())
-        , TypeName(typeName)
-    {
-        pDocName = Object.getDocumentName().c_str();
-        pObjectName = Object.getObjectName().c_str();
-        pSubName = Object.getSubName().c_str();
-        pTypeName = TypeName.c_str();
-    }
-
-    SelectionChanges(const SelectionChanges& other)
-    {
-        *this = other;
-    }
-
-    SelectionChanges& operator=(const SelectionChanges& other)
-    {
-        Type = other.Type;
-        SubType = other.SubType;
-        x = other.x;
-        y = other.y;
-        z = other.z;
-        Object = other.Object;
-        TypeName = other.TypeName;
-        pDocName = Object.getDocumentName().c_str();
-        pObjectName = Object.getObjectName().c_str();
-        pSubName = Object.getSubName().c_str();
-        pTypeName = TypeName.c_str();
-        pOriginalMsg = other.pOriginalMsg;
-        return *this;
-    }
-
-    SelectionChanges(SelectionChanges&& other)
-    {
-        *this = std::move(other);
-    }
-
-    SelectionChanges& operator=(SelectionChanges&& other)
-    {
-        Type = other.Type;
-        SubType = other.SubType;
-        x = other.x;
-        y = other.y;
-        z = other.z;
-        Object = std::move(other.Object);
-        TypeName = std::move(other.TypeName);
-        pDocName = Object.getDocumentName().c_str();
-        pObjectName = Object.getObjectName().c_str();
-        pSubName = Object.getSubName().c_str();
-        pTypeName = TypeName.c_str();
-        pOriginalMsg = other.pOriginalMsg;
-        return *this;
-    }
-
-    MsgType Type;
-    MsgSource SubType;
-
-    const char* pDocName;
-    const char* pObjectName;
-    const char* pSubName;
-    const char* pTypeName;
-    float x;
-    float y;
-    float z;
-
-    App::SubObjectT Object;
-    std::string TypeName;
-
-    // Original selection message in case resolve!=0
-    const SelectionChanges* pOriginalMsg = nullptr;
-};
-
-class ViewProviderDocumentObject;
-
-/**
- * The SelectionObserver class simplifies the step to write classes that listen
- * to what happens to the selection.
- *
- * @author Werner Mayer
- */
-class GuiExport SelectionObserver
-{
-
-public:
-    /** Constructor
-     *
-     * @param attach: whether to   attach this observer on construction
-     * @param resolve: sub-object resolving mode.
-     */
-    explicit SelectionObserver(bool attach = true, ResolveMode resolve = ResolveMode::OldStyleElement);
-    /** Constructor
-     *
-     * @param vp: filtering view object.
-     * @param attach: whether to attach this observer on construction
-     * @param resolve: sub-object resolving mode.
-     *
-     * Constructs an selection observer that receives only selection event of
-     * objects within the same document as the input view object.
-     */
-    explicit SelectionObserver(
-        const Gui::ViewProviderDocumentObject* vp,
-        bool attach = true,
-        ResolveMode resolve = ResolveMode::OldStyleElement
-    );
-
-    virtual ~SelectionObserver();
-    bool blockSelection(bool block);
-    bool isSelectionBlocked() const;
-    bool isSelectionAttached() const;
-
-    /** Attaches to the selection. */
-    void attachSelection();
-    /** Detaches from the selection. */
-    void detachSelection();
-    /** clears the document scope filter, allowing cross-document selection events. */
-    void clearDocumentScope()
-    {
-        documentScopeName.clear();
-    }
-
-private:
-    virtual void onSelectionChanged(const SelectionChanges& msg) = 0;
-    void _onSelectionChanged(const SelectionChanges& msg);
-
-private:
-    using Connection = fastsignals::connection;
-    Connection connectSelection;
-    std::string filterDocName;
-    std::string filterObjName;
-    ResolveMode resolve;
-    std::string documentScopeName;
-    bool blockedSelection;
-};
-
-/** SelectionGate
- * The selection gate allows or disallows selection of certain types.
- * It has to be registered to the selection.
- */
-class GuiExport SelectionGate
-{
-public:
-    virtual ~SelectionGate() = default;
-    virtual bool allow(App::Document*, App::DocumentObject*, const char*) = 0;
-    /** @brief filter all available types
-     *  @param allTypesForGeometry Every type available to select (ex. {"Vertex", "Edge"})
-     *  @returns a set of filtered types (ex. {"Vertex"})
-     */
-    virtual std::unordered_set<std::string> getGatedTypes(
-        const std::vector<const char*>& allTypesForGeometry
-    ) const
-    {
-        (void)allTypesForGeometry;
-        return {};
-    }
-
-    /**
-     * @brief notAllowedReason is a string that sets the message to be
-     * displayed in statusbar for cluing the user on why is the selection not
-     * allowed. Set this variable in allow() implementation. Enclose the
-     * literal into QT_TR_NOOP() for translatability.
-     */
-    std::string notAllowedReason;
-};
 
 /** The Selection class
  *  The selection singleton keeps track of the selection state of
@@ -729,37 +479,6 @@ public:
     static SelectionSingleton& instance();
     static void destruct();
     friend class SelectionFilter;
-
-    // Python interface
-    static PyMethodDef Methods[];
-
-protected:
-    static PyObject* sAddSelection(PyObject* self, PyObject* args);
-    static PyObject* sUpdateSelection(PyObject* self, PyObject* args);
-    static PyObject* sRemoveSelection(PyObject* self, PyObject* args);
-    static PyObject* sClearSelection(PyObject* self, PyObject* args);
-    static PyObject* sIsSelected(PyObject* self, PyObject* args);
-    static PyObject* sCountObjectsOfType(PyObject* self, PyObject* args);
-    static PyObject* sGetSelection(PyObject* self, PyObject* args);
-    static PyObject* sSetPreselection(PyObject* self, PyObject* args, PyObject* kwd);
-    static PyObject* sGetPreselection(PyObject* self, PyObject* args);
-    static PyObject* sRemPreselection(PyObject* self, PyObject* args);
-    static PyObject* sGetCompleteSelection(PyObject* self, PyObject* args);
-    static PyObject* sGetSelectionEx(PyObject* self, PyObject* args);
-    static PyObject* sGetSelectionObject(PyObject* self, PyObject* args);
-    static PyObject* sSetSelectionStyle(PyObject* self, PyObject* args);
-    static PyObject* sAddSelObserver(PyObject* self, PyObject* args);
-    static PyObject* sRemSelObserver(PyObject* self, PyObject* args);
-    static PyObject* sAddSelectionGate(PyObject* self, PyObject* args);
-    static PyObject* sRemoveSelectionGate(PyObject* self, PyObject* args);
-    static PyObject* sGetPickedList(PyObject* self, PyObject* args);
-    static PyObject* sEnablePickedList(PyObject* self, PyObject* args);
-    static PyObject* sPreselect(PyObject* self, PyObject* args);
-    static PyObject* sSetVisible(PyObject* self, PyObject* args);
-    static PyObject* sPushSelStack(PyObject* self, PyObject* args);
-    static PyObject* sHasSelection(PyObject* self, PyObject* args);
-    static PyObject* sHasSubSelection(PyObject* self, PyObject* args);
-    static PyObject* sGetSelectionFromStack(PyObject* self, PyObject* args);
 
 protected:
     /// Construction
